+++ lib/src/http/http.dart
@@ -1,3 +1,4 @@
+/// The parent HTTP library that exposes all the constituting libraries.
 library jaguar.src.http;
 
 export 'context/context.dart';
@@ -7,5 +8,6 @@ export 'params/import.dart';
 export 'reflect/reflect.dart';
 export 'request/request.dart';
 export 'response/import.dart';
+export 'route_chain/route_chain.dart';
 export 'session/session.dart';
 export 'websocket/websocket.dart';