+++ lib/src/http/reflect/reflect.dart
@@ -1,25 +1,38 @@
 library jaguar.http.reflect;
 
 import 'dart:async';
-import 'package:jaguar/jaguar.dart' as j;
+import 'package:jaguar/jaguar.dart';
 import 'dart:mirrors';
+import 'dart:collection';
 
-part 'route.dart';
+/// Reflects the given [api] and returns an instance of [ReflectedApi]
+ReflectedApi reflectApi(Object api) => new ReflectedApi(api);
 
-class JaguarReflected implements j.RequestHandler {
-  final dynamic _handler;
+/// Composing routes are parsed by reflecting the provided [_api]. Alternative
+/// is to to source generate.
+class ReflectedApi implements RequestHandler {
+  /// The api instance being reflected
+  final dynamic _api;
 
-  final _routes = <ReflectedRoute>[];
+  /// Routes parsed from [_api]
+  final _routes = <RouteChain>[];
 
-  JaguarReflected(this._handler) {
+  /// The composing routes.
+  ///
+  /// Publicly exposes an immutable version of [_routes]
+  UnmodifiableListView<RouteChain> get routes =>
+      new UnmodifiableListView<RouteChain>(_routes);
+
+  /// Constructs a
+  ReflectedApi(this._api) {
     _parseApi();
   }
 
-  Future<j.Response> handleRequest(j.Context ctx, {String prefix: ''}) async {
-    for (ReflectedRoute route in _routes) {
-      j.Response response =
-          await route.handleRequest(ctx, prefix: prefix + route.prefix);
-      if (response is j.Response) {
+  /// Handles requests
+  Future<Response> handleRequest(Context ctx, {String prefix: ''}) async {
+    for (RouteChain route in _routes) {
+      Response response = await route.handleRequest(ctx, prefix: prefix);
+      if (response is Response) {
         return response;
       }
     }
@@ -28,21 +41,21 @@ class JaguarReflected implements j.RequestHandler {
   }
 
   void _parseApi() {
-    final InstanceMirror im = reflect(_handler);
-    final List<j.Api> apis = im.type.metadata
+    final InstanceMirror im = reflect(_api);
+    final List<Api> apis = im.type.metadata
         .map((InstanceMirror aim) => aim.reflectee)
-        .where((dynamic ref) => ref is j.Api)
-        .toList() as List<j.Api>;
+        .where((dynamic ref) => ref is Api)
+        .toList() as List<Api>;
     if (apis.length == 0) {
       throw new Exception('Handler is not decorated with Api!');
     }
 
     final List<Symbol> wrappers = _detectWrappers(im.type.metadata);
 
-    final List<j.ExceptionHandler> exceptionHandlers = im.type.metadata
-        .where((InstanceMirror annot) => annot.reflectee is j.ExceptionHandler)
+    final List<ExceptionHandler> exceptionHandlers = im.type.metadata
+        .where((InstanceMirror annot) => annot.reflectee is ExceptionHandler)
         .map((InstanceMirror annot) => annot.reflectee)
-        .toList() as List<j.ExceptionHandler>;
+        .toList() as List<ExceptionHandler>;
 
     _parse(im, apis.first.path,
         topWrapper: wrappers, topExceptionHandlers: exceptionHandlers);
@@ -50,16 +63,16 @@ class JaguarReflected implements j.RequestHandler {
 
   void _parse(InstanceMirror im, String pathPrefix,
       {List<Symbol> topWrapper: const [],
-      List<j.ExceptionHandler> topExceptionHandlers: const []}) {
+      List<ExceptionHandler> topExceptionHandlers: const []}) {
     im.type.declarations.forEach((Symbol s, DeclarationMirror decl) {
       if (decl.isPrivate) return;
 
       // Collect IncludeApi
       if (decl is VariableMirror) {
-        final List<j.IncludeApi> groups = decl.metadata
-            .where((InstanceMirror annot) => annot.reflectee is j.IncludeApi)
+        final List<IncludeApi> groups = decl.metadata
+            .where((InstanceMirror annot) => annot.reflectee is IncludeApi)
             .map((InstanceMirror annot) => annot.reflectee)
-            .toList() as List<j.IncludeApi>;
+            .toList() as List<IncludeApi>;
 
         if (groups.length == 0) return;
 
@@ -72,10 +85,10 @@ class JaguarReflected implements j.RequestHandler {
           throw new Exception('Group cannot be null!');
         }
 
-        List<j.Api> rg = gim.type.metadata
-            .where((InstanceMirror annot) => annot.reflectee is j.Api)
+        List<Api> rg = gim.type.metadata
+            .where((InstanceMirror annot) => annot.reflectee is Api)
             .map((InstanceMirror annot) => annot.reflectee)
-            .toList() as List<j.Api>;
+            .toList() as List<Api>;
 
         if (rg.length == 0) {
           throw new Exception('Included API must be annotated with Api!');
@@ -91,10 +104,10 @@ class JaguarReflected implements j.RequestHandler {
 
       if (decl is! MethodMirror) return;
 
-      final List<j.RouteBase> routes = decl.metadata
-          .where((InstanceMirror annot) => annot.reflectee is j.RouteBase)
+      final List<RouteBase> routes = decl.metadata
+          .where((InstanceMirror annot) => annot.reflectee is RouteBase)
           .map((InstanceMirror annot) => annot.reflectee)
-          .toList() as List<j.RouteBase>;
+          .toList() as List<RouteBase>;
       if (routes.length == 0) return;
 
       final List<Symbol> wrappers = _detectWrappers(decl.metadata);
@@ -103,17 +116,16 @@ class JaguarReflected implements j.RequestHandler {
 
       // Collect exception handlers
 
-      List<j.ExceptionHandler> exceptionHandlers = decl.metadata
-          .where(
-              (InstanceMirror annot) => annot.reflectee is j.ExceptionHandler)
+      List<ExceptionHandler> exceptionHandlers = decl.metadata
+          .where((InstanceMirror annot) => annot.reflectee is ExceptionHandler)
           .map((InstanceMirror annot) => annot.reflectee)
-          .toList() as List<j.ExceptionHandler>;
+          .toList() as List<ExceptionHandler>;
       exceptionHandlers.insertAll(0, topExceptionHandlers);
 
       InstanceMirror method = im.getField(s);
 
       routes
-          .map((j.RouteBase route) => new ReflectedRoute.buildForClass(
+          .map((RouteBase route) => new RouteChain.reflectClass(
               method.reflectee,
               route,
               pathPrefix,
@@ -125,17 +137,14 @@ class JaguarReflected implements j.RequestHandler {
   }
 }
 
-JaguarReflected reflectJaguar(Object routeGroup) =>
-    new JaguarReflected(routeGroup);
-
 List<Symbol> _detectWrappers(List<InstanceMirror> annots) {
   final wrappers = <Symbol>[];
 
   for (InstanceMirror annot in annots) {
     dynamic ref = annot.reflectee;
-    if (ref is j.WrapOne) {
+    if (ref is WrapOne) {
       wrappers.add(ref.interceptor);
-    } else if (ref is j.Wrap) {
+    } else if (ref is Wrap) {
       wrappers.addAll(ref.interceptors);
     }
   }